SIG: Move AS map to egress
The map is part of egress, so move it there.

This is part of the effort to minimize the interface between main and
egress.

Also renames "core" package to "asmap".
