+++ go/sig/BUILD.bazel
@@ -12,10 +12,10 @@ go_library(
         "//go/lib/fatal:go_default_library",
         "//go/lib/log:go_default_library",
         "//go/sig/base:go_default_library",
-        "//go/sig/base/core:go_default_library",
         "//go/sig/config:go_default_library",
         "//go/sig/disp:go_default_library",
         "//go/sig/egress:go_default_library",
+        "//go/sig/egress/asmap:go_default_library",
         "//go/sig/egress/reader:go_default_library",
         "//go/sig/ingress:go_default_library",
         "//go/sig/internal/sigconfig:go_default_library",
