+++ go/sig/main.go
@@ -32,10 +32,10 @@ import (
 	"github.com/scionproto/scion/go/lib/fatal"
 	"github.com/scionproto/scion/go/lib/log"
 	"github.com/scionproto/scion/go/sig/base"
-	"github.com/scionproto/scion/go/sig/base/core"
 	"github.com/scionproto/scion/go/sig/config"
 	"github.com/scionproto/scion/go/sig/disp"
 	"github.com/scionproto/scion/go/sig/egress"
+	"github.com/scionproto/scion/go/sig/egress/asmap"
 	"github.com/scionproto/scion/go/sig/egress/reader"
 	"github.com/scionproto/scion/go/sig/ingress"
 	"github.com/scionproto/scion/go/sig/internal/sigconfig"
@@ -216,7 +216,7 @@ func loadConfig(path string) bool {
 		log.Error("loadConfig: Failed", "err", err)
 		return false
 	}
-	ok := core.Map.ReloadConfig(cfg)
+	ok := asmap.Map.ReloadConfig(cfg)
 	if !ok {
 		return false
 	}
